Added tests for Project specific mappings in ColumnSwapper
