+++ Rdmp.Core.Tests/DataLoad/Engine/Integration/PipelineTests/Components/ColumnSwapperTests.cs
@@ -11,8 +11,11 @@ using System.Linq;
 using FAnsi;
 using FAnsi.Discovery;
 using FAnsi.Extensions;
+using Moq;
 using NUnit.Framework;
 using Rdmp.Core.Curation.Data;
+using Rdmp.Core.DataExport.Data;
+using Rdmp.Core.DataExport.DataExtraction.Commands;
 using Rdmp.Core.DataFlowPipeline;
 using Rdmp.Core.DataLoad.Modules.DataFlowOperations.Aliases;
 using Rdmp.Core.DataLoad.Modules.DataFlowOperations.Aliases.Exceptions;
@@ -309,9 +312,56 @@ namespace Rdmp.Core.Tests.DataLoad.Engine.Integration.PipelineTests.Components
                 AreBasicallyEquals(1, resultDt.Rows[0]["Out"]);
                 Assert.AreEqual("Dave", resultDt.Rows[0]["Name"]);
             }
-
         }
-        
+
+        [Test]
+        public void TestColumnSwapper_ProjectSpecificMappings()
+        {
+            var dt = new DataTable();
+            dt.Columns.Add("In");
+            dt.Columns.Add("Out");
+            dt.Columns.Add("Proj");
+
+            //Anonymise A and B differently depending on ProjectNumber (valid project numbers are 1 and 2)
+            dt.Rows.Add("A", 1,1);
+            dt.Rows.Add("A", 2,2);
+            dt.Rows.Add("B", 3,1);
+            dt.Rows.Add("B", 4,2);
+
+            var db = GetCleanedServer(DatabaseType.MicrosoftSQLServer);
+
+            Import(db.CreateTable("Map", dt), out var map, out var mapCols);
+
+            var swapper = new ColumnSwapper();
+            swapper.MappingFromColumn = mapCols.Single(c => c.GetRuntimeName().Equals("In"));
+            swapper.MappingToColumn = mapCols.Single(c => c.GetRuntimeName().Equals("Out"));
+            swapper.WHERELogic = "Proj = $n";
+            
+            // initialize with a mock that returns ProjectNumber 1
+            swapper.PreInitialize(GetMockExtractDatasetCommand(), new ThrowImmediatelyDataLoadEventListener());
+
+            swapper.Check(new ThrowImmediatelyCheckNotifier());
+
+            var dtToSwap = new DataTable();
+
+            dtToSwap.Columns.Add("In");
+            dtToSwap.Columns.Add("Name");
+            dtToSwap.Columns.Add("Age");
+
+            dtToSwap.Rows.Add("A", "Dave", 30);
+            dtToSwap.Rows.Add("B", "Frank", 50);
+
+            var resultDt = swapper.ProcessPipelineData(dtToSwap, new ThrowImmediatelyDataLoadEventListener(), new GracefulCancellationToken());
+
+            Assert.AreEqual(2, resultDt.Rows.Count);
+
+            // Should have project specific results for A of 1 and for B of 3 because the ProjectNumber is 1
+            AreBasicallyEquals(1, resultDt.Rows[0]["Out"]);
+            Assert.AreEqual("Dave", resultDt.Rows[0]["Name"]);
+            AreBasicallyEquals(3, resultDt.Rows[1]["Out"]);
+            Assert.AreEqual("Frank", resultDt.Rows[1]["Name"]);
+        }
+
         /// <summary>
         /// Tests ColumnSwapper when there are null values in the input <see cref="DataTable"/> being processed
         /// </summary>
@@ -484,6 +534,27 @@ namespace Rdmp.Core.Tests.DataLoad.Engine.Integration.PipelineTests.Components
             Assert.AreEqual(1, resultDt.Rows.Count);
             AreBasicallyEquals(2, resultDt.Rows[0]["Out"]);
             Assert.AreEqual("Dave", resultDt.Rows[0]["Name"]);
+        }
+
+        private IExtractDatasetCommand GetMockExtractDatasetCommand()
+        {
+            var mockPj = Mock.Of<IProject>(p =>
+                p.Name == "My Project" &&
+                p.ProjectNumber == 1
+            );
+
+            var mockConfig = Mock.Of<IExtractionConfiguration>();
+
+            var mockSelectedDatasets = Mock.Of<ISelectedDataSets>(sds =>
+                sds.ExtractionConfiguration == mockConfig
+            );
+
+            var mockExtractDsCmd = Mock.Of<IExtractDatasetCommand>(d =>
+                d.Project == mockPj &&
+                d.SelectedDataSets == mockSelectedDatasets
+            );
+
+            return mockExtractDsCmd;
         }
     }
 }
